Change key length data type to short

+++ src/radixtree.c
@@ -218,7 +218,7 @@ DATA_FOUND:
 /**
  * Add new child node after given node
  */
-static Node *_build_node(Node *node, unsigned char *string, unsigned int length)
+static Node *_build_node(Node *node, unsigned char *string, unsigned short length)
 {
 	if (length >= 1) {
 		node = bsearch_insert(node, string[0]);
@@ -355,7 +355,7 @@ static void _pluck_node(Node *node, ScanStatus *status, ScanMetadata *meta)
 	}
 }
 
-void *radix_tree_get(Node *tree, unsigned char *string, unsigned int length)
+void *radix_tree_get(Node *tree, unsigned char *string, unsigned short length)
 {
 	trace("RADIXTREE-GET(%p)", tree);
 
@@ -373,7 +373,7 @@ void *radix_tree_get(Node *tree, unsigned char *string, unsigned int length)
 	}
 }
 
-static Node *_build_data_node(Node *tree, unsigned char *string, unsigned int length)
+static Node *_build_data_node(Node *tree, unsigned char *string, unsigned short length)
 {
 	Node *data_node;
 
@@ -392,7 +392,7 @@ static Node *_build_data_node(Node *tree, unsigned char *string, unsigned int le
 	return data_node;
 }
 
-void radix_tree_set(Node *tree, unsigned char *string, unsigned int length, void *data)
+void radix_tree_set(Node *tree, unsigned char *string, unsigned short length, void *data)
 {
 	trace("RADIXTREE-SET(%p)", tree);
 
@@ -400,7 +400,7 @@ void radix_tree_set(Node *tree, unsigned char *string, unsigned int length, void
 	data_node->data = data;
 }
 
-int radix_tree_contains(Node *tree, unsigned char *string, unsigned int length)
+int radix_tree_contains(Node *tree, unsigned char *string, unsigned short length)
 {
 	trace("RADIXTREE-CONTAINS(%p)", tree);
 
@@ -418,7 +418,7 @@ int radix_tree_contains(Node *tree, unsigned char *string, unsigned int length)
 	}
 }
 
-void *radix_tree_try_set(Node *tree, unsigned char *string, unsigned int length, void *data)
+void *radix_tree_try_set(Node *tree, unsigned char *string, unsigned short length, void *data)
 {
 	trace("RADIXTREE-TRY-SET(%p)", tree);
 
@@ -431,7 +431,7 @@ void *radix_tree_try_set(Node *tree, unsigned char *string, unsigned int length,
 	return previous_data;
 }
 
-void radix_tree_remove(Node *tree, unsigned char *string, unsigned int length)
+void radix_tree_remove(Node *tree, unsigned char *string, unsigned short length)
 {
 	trace("RADIXTREE-REMOVE(%p)", tree);
 
@@ -451,7 +451,7 @@ void radix_tree_remove(Node *tree, unsigned char *string, unsigned int length)
 	}
 }
 
-void init_status(ScanStatus *status, ScanStatus *poststatus, unsigned char *string, unsigned int length)
+void init_status(ScanStatus *status, ScanStatus *poststatus, unsigned char *string, unsigned short length)
 {
 	status->index = 0;
 	status->subindex = 0;
@@ -470,7 +470,7 @@ void init_status(ScanStatus *status, ScanStatus *poststatus, unsigned char *stri
 	poststatus->found = 0;
 }
 
-void **radix_tree_get_next(Node *tree, unsigned char *string, unsigned int length)
+void **radix_tree_get_next(Node *tree, unsigned char *string, unsigned short length)
 {
 	Node *res;
 	ScanStatus status;
